Improved logic for swapping inserted-navigation flags.

+++ classes/Import/Importer.php
@@ -14,6 +14,7 @@ namespace OpenLab\Modules\Import;
 use WP_Error;
 use XMLReader;
 use OpenLab\Modules\Logger\Logger;
+use OpenLab\Modules\Module;
 
 /**
  * Importer class.
@@ -1961,9 +1962,9 @@ class Importer {
 			$this->post_process_comments( $this->requires_remapping['comment'] );
 		}
 
+		$this->post_process_cpt_tax_map();
 		$this->post_process_block_attributes();
 		$this->post_process_post_meta();
-		$this->post_process_cpt_tax_map();
 	}
 
 	/**
@@ -2282,13 +2283,19 @@ class Importer {
 
 		// First, loop through and change the inserted_navigation keys to avoid dupes.
 		foreach ( $mapping as $old_post_id => $new_post_id ) {
-			$old_meta_key = 'openlab_modules_inserted_navigation_' . (string) $old_post_id;
-			$new_meta_key = 'openlab_modules_inserted_navigation_' . (string) $new_post_id;
+			$post_module_ids = Module::get_module_ids_of_page( $new_post_id );
+			foreach ( $post_module_ids as $new_post_module_id ) {
+				$old_post_module_id = array_search( $new_post_module_id, $mapping, true );
+				if ( $old_post_module_id ) {
+					$old_meta_key = 'openlab_modules_inserted_navigation_' . (string) $old_post_module_id;
+					$new_meta_key = 'openlab_modules_inserted_navigation_' . (string) $new_post_module_id;
 
-			$old_meta_value = get_post_meta( $new_post_id, $old_meta_key, true );
-			if ( $old_meta_value ) {
-				delete_post_meta( $new_post_id, $old_meta_key );
-				update_post_meta( $new_post_id, $new_meta_key, '1' );
+					$old_meta_value = get_post_meta( $new_post_id, $old_meta_key, true );
+					if ( $old_meta_value ) {
+						delete_post_meta( $new_post_id, $old_meta_key );
+						update_post_meta( $new_post_id, $new_meta_key, '1' );
+					}
+				}
 			}
 		}
